Join using QrCode Image
